Move back connection definition to the terminal package
