+++ packages/terminal/package.json
@@ -14,6 +14,7 @@
   },
   "theiaExtensions": [
     {
+      "frontend": "lib/browser/terminal-frontend-module",
       "backend": "lib/node/terminal-backend-module"
     }
   ],
